requirements.txt
python-telegram-bot==20.0a
configparser
oauth2
peewee
instaloader
beautifulsoup4
telegraph[aio]